Make logic more abstract and closer to root to allow for d&d
