+++ components/ActivityForm.jsx
@@ -1,7 +1,9 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { v4 as uuid } from 'uuid';
 
-const ActivityForm = ({ activity, onSave, setFormShowing }) => {
+const ActivityForm = ({ activity, day, onSave, setFormShowing }) => {
+  console.log(day);
+
   const fieldIDs = {
     title: uuid(),
     location: uuid(),
@@ -28,7 +30,7 @@ const ActivityForm = ({ activity, onSave, setFormShowing }) => {
 
   const handleSave = (e) => {
     e.preventDefault();
-    onSave(editedActivity);
+    onSave(day, editedActivity);
 
     setEditedActivity({
       title: '',
